refactor(eval): consolidate evaluators into single parametrized test
- Remove per-query evaluator classes and registry
- Replace with single `AgentEvaluator` that accepts `query_id`
- Move resolution assertions to module-level dict indexed by query_id
- Update conftest to parametrize across all queries instead of single query
- Rename `evaluators/` to `evaluations/` and consolidate into `agent_evaluator.py`
- Replace `test_evaluate_query_*.py` with single `test_evaluate_agent.py`
- Remove `run`,
