+++ gen-ai/examples/evaluations/data_analyst_agent_evaluation/Makefile
@@ -1,20 +1,11 @@
 PACKAGE    := $(shell basename $(CURDIR))
 
-.PHONY: install run run-query run-question test test-query test-question
+.PHONY: install test test-query test-question
 
 install:
 	export UV_INDEX_GEN_AI_INTERNAL_USERNAME=oauth2accesstoken && \
 	export UV_INDEX_GEN_AI_INTERNAL_PASSWORD="$$(gcloud auth print-access-token)" && \
 	uv sync --all-extras
-
-run:
-	python -m $(PACKAGE).runner
-
-run-query:
-	python -m $(PACKAGE).runner --query-id $(IDS)
-
-run-question:
-	python -m $(PACKAGE).runner --question-id $(IDS)
 
 test:
 	pytest
