+++ gen-ai/examples/evaluations/data_analyst_agent_evaluation/evaluations/__init__.py
@@ -1,0 +1,3 @@
+"""Agent evaluation classes."""
+
+from .agent_evaluator import AgentEvaluator
