+++ gen-ai/examples/evaluations/data_analyst_agent_evaluation/pyproject.toml
@@ -22,5 +22,5 @@ gllm-evals = { index = "gen-ai-internal" }
 
 [tool.pytest.ini_options]
 pythonpath = ["."]
-testpaths = ["tests"]
+testpaths = ["evaluations"]
 addopts = "-v"
